Edición de perfil - #151

+++ youroom/youroom/base_tests.py
@@ -17,15 +17,16 @@ class BaseTestCase(APITestCase):
         self.u.email = 'prueba@gmail.com'
         self.u.isActive = True
         self.u.save()
-        self.p = UsuarioPerfil.objects.get_or_create(user=self.u, totalPuntos=100)[0]
+        self.p = UsuarioPerfil.objects.get_or_create(user=self.u, totalPuntos=100, descripcion ='descripcion prueba')[0]
         self.c = ContadorVida.objects.get_or_create(perfil=self.p, estaActivo=True)[0]
         call_command('loaddata', 'products.json', verbosity=0)
 
     def tearDown(self):
         self.client = None
-        filelist = [f for f in os.listdir('./static/media/publicaciones/') if f.endswith(".png")]
-        for f in filelist:
-            os.remove(os.path.join('./static/media/publicaciones/', f))
+        if os.path.exists('./static/media/publicaciones/'):
+            filelist = [f for f in os.listdir('./static/media/publicaciones/') if f.endswith(".png")]
+            for f in filelist:
+                os.remove(os.path.join('./static/media/publicaciones/', f))
 
     def generate_photo_file(self):
         file = io.BytesIO()
